+++ packages/utils/interactive-hooks/src/index.ts
@@ -1,3 +1,6 @@
+export * from './Pressable.types';
+export * from './useAsPressable';
+export * from './useFocusState';
+export * from './useHoverState';
+export * from './usePressState';
 export * from './useViewCommandFocus';
-export * from './Pressable.hooks';
-export * from './Pressable.hooks.helpers';
